Reorganize mocks and fix some linter issues
