+++ aggregatestore/snapshotting_store_test.go
@@ -14,45 +14,6 @@ import (
 	"github.com/gofrs/uuid/v5"
 )
 
-type mockStore2[E estoria.Entity] struct {
-	NewFn     func(id uuid.UUID) (*aggregatestore.Aggregate[E], error)
-	LoadFn    func(context.Context, typeid.UUID, aggregatestore.LoadOptions) (*aggregatestore.Aggregate[E], error)
-	HydrateFn func(context.Context, *aggregatestore.Aggregate[E], aggregatestore.HydrateOptions) error
-	SaveFn    func(context.Context, *aggregatestore.Aggregate[E], aggregatestore.SaveOptions) error
-}
-
-func (s *mockStore2[E]) New(id uuid.UUID) (*aggregatestore.Aggregate[E], error) {
-	if s.NewFn != nil {
-		return s.NewFn(id)
-	}
-
-	return nil, fmt.Errorf("unexpected call: New(id=%s)", id.String())
-}
-
-func (s *mockStore2[E]) Load(ctx context.Context, aggregateID typeid.UUID, opts aggregatestore.LoadOptions) (*aggregatestore.Aggregate[E], error) {
-	if s.LoadFn != nil {
-		return s.LoadFn(ctx, aggregateID, opts)
-	}
-
-	return nil, fmt.Errorf("unexpected call: Load(aggregateID=%s, opts=%v)", aggregateID, opts)
-}
-
-func (s *mockStore2[E]) Hydrate(ctx context.Context, aggregate *aggregatestore.Aggregate[E], opts aggregatestore.HydrateOptions) error {
-	if s.HydrateFn != nil {
-		return s.HydrateFn(ctx, aggregate, opts)
-	}
-
-	return fmt.Errorf("unexpected call: Hydrate(aggregate=%v, opts=%v)", aggregate, opts)
-}
-
-func (s *mockStore2[E]) Save(ctx context.Context, aggregate *aggregatestore.Aggregate[E], opts aggregatestore.SaveOptions) error {
-	if s.SaveFn != nil {
-		return s.SaveFn(ctx, aggregate, opts)
-	}
-
-	return fmt.Errorf("unexpected call: Save(aggregate=%v, opts=%v)", aggregate, opts)
-}
-
 type mockSnapshotStore struct {
 	ReadSnapshotFn  func(context.Context, typeid.UUID, snapshotstore.ReadSnapshotOptions) (*snapshotstore.AggregateSnapshot, error)
 	WriteSnapshotFn func(context.Context, *snapshotstore.AggregateSnapshot) error
@@ -618,7 +579,7 @@ func TestSnapshottingStore_Hydrate(t *testing.T) {
 			},
 			haveSnapshotStore: &mockSnapshotStore{
 				ReadSnapshotFn: func(_ context.Context, _ typeid.UUID, _ snapshotstore.ReadSnapshotOptions) (*snapshotstore.AggregateSnapshot, error) {
-					return nil, nil
+					return nil, snapshotstore.ErrSnapshotNotFound
 				},
 			},
 			haveAggregate: func() *aggregatestore.Aggregate[*mockEntity] {
